// src/app/shared/classes/ignoreInterceptorType.enum.ts
export enum IgnoreInterceptorTypeNames {
    ERROR_302 = 'ignoreInterceptorError302',
    ERROR_503 = 'ignoreInterceptorError503',
    ERROR_500 = 'ignoreInterceptorError500',
    ERROR_404 = 'ignoreInterceptorError404',
    ERROR_403 = 'ignoreInterceptorError403',
    ERROR_401 = 'ignoreInterceptorError401',
    ERROR_400 = 'ignoreInterceptorError400',
}

export namespace IgnoreInterceptorType {
}